silence a couple of more warnings

+++ src/game/AuctionHouseBot/AuctionHouseBot.cpp
@@ -178,13 +178,13 @@ void AuctionHouseBot::Update()
             int32 maxTemplates = m_professionItemsConfig[0] < 0 ? urand(0, m_professionItemsConfig[1] - m_professionItemsConfig[0]) + m_professionItemsConfig[0] : urand(m_professionItemsConfig[0], m_professionItemsConfig[1]);
             if (maxTemplates > 0)
             {
-                for (uint32 templateCounter = 0; templateCounter < maxTemplates; ++templateCounter)
+                for (int32 templateCounter = 0; templateCounter < maxTemplates; ++templateCounter)
                 {
                     uint32 item = m_professionItems[urand(0, m_professionItems.size() - 1)];
                     ItemPrototype const* prototype = ObjectMgr::GetItemPrototype(item);
                     if (!prototype || prototype->Quality == 0 || urand(0, (1 << (prototype->Quality - 1)) - 1) > 0)
                         continue; // make it decreasingly likely that crafted items of higher quality is added to the auction house (white: 100%, green: 50%, blue: 25%, purple: 12.5%, ...)
-                    uint32 count = (uint32) round(prototype->GetMaxStackSize() * urand(m_professionItemsConfig[2], m_professionItemsConfig[3]) / 100.0);
+                    uint32 count = (uint32) round((uint64)prototype->GetMaxStackSize() * urand(m_professionItemsConfig[2], m_professionItemsConfig[3]) / 100.0);
                     itemMap[item] += count;
                 }
             }
@@ -460,7 +460,7 @@ void AuctionHouseBot::AddLootToItemMap(LootStore* store, std::vector<int32>& loo
     int32 maxTemplates = lootConfig[0] < 0 ? urand(0, lootConfig[1] - lootConfig[0]) + lootConfig[0] : urand(lootConfig[0], lootConfig[1]);
     if (maxTemplates <= 0)
         return;
-    for (uint32 templateCounter = 0; templateCounter < maxTemplates; ++templateCounter)
+    for (int32 templateCounter = 0; templateCounter < maxTemplates; ++templateCounter)
     {
         uint32 lootTemplate = urand(0, lootTemplates.size() - 1);
         LootTemplate const* lootTable = store->GetLootFor(lootTemplates[lootTemplate]);
